Created a sidebar ui for changing the depth of trees. Also confirmed with WSL Ubuntu and -fsanitize=leak that there doesn't seem to be any memory leaks. As a side note, running the game on WSL Ubuntu almost works perfectly, but the tree shader is wrong.

+++ include/button.cpp
@@ -2,74 +2,74 @@
 #include "button.hpp"
 
 Button::Button(Button::Owner* owner, Vector2 pos, Vector2 dim, std::string text, 
-    std::function<void(Button&)> on_hover,
-    std::function<void(Button&)> on_hit) {
-    state.hovered = false;
-    state.last_hit = false;
-    state.hit = false;
-    state.owner = owner;
-    
-    std::cout << "created button\n";
-    state.pos = pos;
-    state.dim = dim;
-    state.text = text;
-    this->on_hover = on_hover;
-    this->on_hit = on_hit;
+	std::function<void(Button&)> on_hover,
+	std::function<void(Button&)> on_hit) {
+	state.hovered = false;
+	state.last_hit = false;
+	state.hit = false;
+	state.owner = owner;
+	
+	std::cout << "created button\n";
+	state.pos = pos;
+	state.dim = dim;
+	state.text = text;
+	this->on_hover = on_hover;
+	this->on_hit = on_hit;
 
-    // Default values
-    state.background_color = YELLOW;
-    state.background_color.a = 100;
-    state.text_color = BLACK;
+	// Default values
+	state.background_color = YELLOW;
+	state.background_color.a = 100;
+	state.text_color = BLACK;
 
-    idle_state = state;
+	idle_state = state;
 }
 
 Button::~Button() {
-    std::cout << "deinit button\n";
+	std::cout << "deinit button\n";
 }
 
 void Button::take_input(Vector2 cursor) {
-    auto& pos = state.pos;
-    auto& hit = state.hit;
-    auto& last_hit = state.last_hit;
-    auto& dim = state.dim;
-    auto& hovered = state.hovered;
+	auto& pos = state.pos;
+	auto& hit = state.hit;
+	auto& last_hit = state.last_hit;
+	auto& dim = state.dim;
+	auto& hovered = state.hovered;
 
-    bool horz = pos.x < cursor.x && cursor.x < pos.x + dim.x;
-    bool vert = pos.y < cursor.y && cursor.y < pos.y + dim.y;
-    bool new_hovered = horz && vert;
+	bool horz = pos.x < cursor.x && cursor.x < pos.x + dim.x;
+	bool vert = pos.y < cursor.y && cursor.y < pos.y + dim.y;
+	bool new_hovered = horz && vert;
 
-    if (new_hovered && !hovered) {
-        // Save idle state, aka this state
-        std::cout << "copy constructor button\n";
-        idle_state = state;
-        on_hover(*this);
-    }
-    if (hit && !last_hit) {
-        on_hit(*this);
-        hit = false;
-    }
-    if (!new_hovered && hovered) {
-        state = idle_state;
-    }
+	if (new_hovered && !hovered) {
+		// Save idle state, aka this state
+		std::cout << "copy constructor button\n";
+		idle_state = state;
+		on_hover(*this);
+	}
+	if (hit && !last_hit) {
+		on_hit(*this);
+		hit = false;
+	}
+	if (!new_hovered && hovered) {
+		state = idle_state;
+	}
 
-    hovered = new_hovered;
-    last_hit = hit;
+	hovered = new_hovered;
+	last_hit = hit;
 }
 
 void Button::render() const {
-    auto& background_color = state.background_color;
-    auto& hovered = state.hovered;
-    auto& dim = state.dim;
-    auto& pos = state.pos;
-    auto& text = state.text;
-    auto& text_color = state.text_color;
+	auto& background_color = state.background_color;
+	auto& hovered = state.hovered;
+	auto& dim = state.dim;
+	auto& pos = state.pos;
+	auto& text = state.text;
+	auto& text_color = state.text_color;
 
-    auto color = background_color;
-    if (hovered) {
-        color = lerp(background_color, BLACK, 0.4);
-    }
-    DrawRectangle((int) pos.x, (int) pos.y, (int) dim.x, (int) dim.y, color);
+	auto color = background_color;
+	if (hovered) {
+		color = lerp(background_color, BLACK, 0.4);
+	}
+	DrawRectangle(pos.x, pos.y, dim.x, dim.y, color);
 
-    DrawText(text.c_str(), (int) pos.x, (int) pos.y, 20, text_color);
+	DrawText(text.c_str(), pos.x, pos.y, 20, text_color);
 }
