+++ include/mylib.cpp
@@ -5,147 +5,147 @@
 #include <format>
 
 Vector2I::Vector2I(Vector2 v) {
-    this->x = v.x;
-    this->y = v.y;
+	this->x = v.x;
+	this->y = v.y;
 }
 
 Vector2I::Vector2I(int x, int y) {
-    this->x = x;
-    this->y = y;
+	this->x = x;
+	this->y = y;
 }
 
 Vector2I::Vector2I() {
-    this->x = 0;
-    this->y = 0;
+	this->x = 0;
+	this->y = 0;
 }
 
 Rand::Rand(int seed) {
-    dist = std::uniform_real_distribution<>(0, 1);
-    set_seed(seed);
+	dist = std::uniform_real_distribution<>(0, 1);
+	set_seed(seed);
 }
 
 void Rand::set_seed(int seed) {
-    this->seed = seed;
-    int_gen = std::mt19937(seed);
+	this->seed = seed;
+	int_gen = std::mt19937(seed);
 }
 
 float Rand::gen(float a, float b) {
-    const float norm = dist(int_gen);
-    return norm * (b - a) + a;
+	const float norm = dist(int_gen);
+	return norm * (b - a) + a;
 }
 
 Vector2 Vector2I::to_vec2() {
-    return { (float) x, (float) y };
+	return { (float) x, (float) y };
 }
 
 void operator += (Vector2& a, const Vector2& b) {
-    a.x += b.x;
-    a.y += b.y;
+	a.x += b.x;
+	a.y += b.y;
 }
 
 Vector2 operator + (const Vector2& a, const Vector2& b) {
-    return { a.x + b.x, a.y + b.y };
+	return { a.x + b.x, a.y + b.y };
 }
 
 Vector2 operator * (const Vector2& v, const float& f) {
-    return { v.x * f, v.y * f };
+	return { v.x * f, v.y * f };
 }
 
 Vector2 operator - (const Vector2& a, const Vector2& b) {
-    return { a.x - b.x, a.y - b.y };
+	return { a.x - b.x, a.y - b.y };
 }
 
 Vector2 operator - (const Vector2& v) {
-    return { -v.x, -v.y };
+	return { -v.x, -v.y };
 }
 
 Vector2 operator / (const Vector2& a, const Vector2& b) {
-    return { a.x / b.x, a.y / b.y };
+	return { a.x / b.x, a.y / b.y };
 }
 
 Vector2 operator / (const Vector2& v, const float& f) {
-    return { v.x / f, v.y / f };
+	return { v.x / f, v.y / f };
 }
 
 Vector2 my_rotate(const Vector2& origin, const Vector2& pt, const float& amt) {
-    const Vector2 out = pt - origin;
-    return Vector2 { 
-        out.x * cos(amt) - out.y * sin(amt),
-        out.x * sin(amt) + out.y * cos(amt)
-    } + origin;
+	const Vector2 out = pt - origin;
+	return Vector2 { 
+		out.x * cos(amt) - out.y * sin(amt),
+		out.x * sin(amt) + out.y * cos(amt)
+	} + origin;
 }
 
 float my_angle(const Vector2& v) {
-    const float ang = atan2f(v.y, v.x);
-    if (ang < 0) {
-        return ang + PI * 2.0;
-    }
-    return ang;
+	const float ang = atan2f(v.y, v.x);
+	if (ang < 0) {
+		return ang + PI * 2.0;
+	}
+	return ang;
 }
 
 float dot(const Vector2& a, const Vector2& b) {
-    return a.x * b.x + a.y * b.y;
+	return a.x * b.x + a.y * b.y;
 }
 
 // Indicates direction to rotate towards, either -1 or 1
 float direction_to_rotate(const Vector2& ahead, const Vector2& mobile) {
-    // Use cross product to figure out if we are on the left or right side
-    return ahead.x * mobile.y - ahead.y * mobile.x < 0 ? 1 : -1;
+	// Use cross product to figure out if we are on the left or right side
+	return ahead.x * mobile.y - ahead.y * mobile.x < 0 ? 1 : -1;
 }
 
 // Does not work past 180 degrees
 float my_angle_from(const Vector2& a, const Vector2& b) {
-    const float cos_theta = dot(a, b) / my_length(a) / my_length(b);
-    return acos(cos_theta);
+	const float cos_theta = dot(a, b) / my_length(a) / my_length(b);
+	return acos(cos_theta);
 }
 
 float my_length(const Vector2& v) {
-    return sqrtf((v.x * v.x) + (v.y * v.y));
+	return sqrtf((v.x * v.x) + (v.y * v.y));
 }
 
 Vector2 my_normalize(const Vector2& v) {
-    const float length = my_length(v);
-    return { v.x / length, v.y / length };
+	const float length = my_length(v);
+	return { v.x / length, v.y / length };
 }
 
 Vector2 perp_rhr(const Vector2& v) {
-    const float vx = -v.y;
-    const float vy = v.x;
-    return { vx, vy };
+	const float vx = -v.y;
+	const float vy = v.x;
+	return { vx, vy };
 }
 
 Vector2 unit_vector(const float& f) {
-    return { cosf(f), sinf(f) };
+	return { cosf(f), sinf(f) };
 }
 
 std::string to_str(const Vector2& v, const int& decimal_pts) {
-    return std::format("({:.{}f}, {:.{}f})", v.x, decimal_pts, v.y, decimal_pts);
+	return std::format("({:.{}f}, {:.{}f})", v.x, decimal_pts, v.y, decimal_pts);
 }
 
 float snap(const float& f, const float& by) {
-    return int(f * by) / by;
+	return (float) int(f * by) / by;
 }
 
 Color lerp(const Color& a, const Color& b, const float& amt) {
-    return {
-        (unsigned char) ((float(b.r) - float(a.r)) * amt + a.r),
-        (unsigned char) ((float(b.g) - float(a.g)) * amt + a.g),
-        (unsigned char) ((float(b.b) - float(a.b)) * amt + a.b),
-        (unsigned char) ((float(b.a) - float(a.a)) * amt + a.a)
-    };
+	return {
+		(unsigned char) ((float(b.r) - float(a.r)) * amt + a.r),
+		(unsigned char) ((float(b.g) - float(a.g)) * amt + a.g),
+		(unsigned char) ((float(b.b) - float(a.b)) * amt + a.b),
+		(unsigned char) ((float(b.a) - float(a.a)) * amt + a.a)
+	};
 }
 
 Texture2D load_dummy_tex() {
-    // Replace with a better dummy texture;
+	// Replace with a better dummy texture;
 
-    // Maybe below is better? 
-    // auto blank = GenImageColor(tree_tex_bounds.x, tree_tex_bounds.y, BLANK);
-    // return LoadTextureFromImage(blank);
-    std::cout << "tree img exists? " << FileExists("assets/tree_texture.png") << "\n";
-    return LoadTexture("assets/tree_texture.png");
+	// Maybe below is better? 
+	// auto blank = GenImageColor(tree_tex_bounds.x, tree_tex_bounds.y, BLANK);
+	// return LoadTextureFromImage(blank);
+	std::cout << "tree img exists? " << FileExists("assets/tree_texture.png") << "\n";
+	return LoadTexture("assets/tree_texture.png");
 }
 
 bool pt_in_rect(const Vector2& pt, const Vector2& pos, const Vector2& dim) {
-    return pos.x <= pt.x && pt.x <= pos.x + dim.x &&
-           pos.y <= pt.y && pt.y <= pos.y + dim.y;
+	return pos.x <= pt.x && pt.x <= pos.x + dim.x &&
+		   pos.y <= pt.y && pt.y <= pos.y + dim.y;
 }
